Rename package to s3.
