+++ solr/contrib/blob-repository/src/java/org/apache/solr/s3/S3BackupRepositoryConfig.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.blob.s3;
+package org.apache.solr.s3;
 
 import com.google.common.base.Strings;
 import org.apache.solr.common.util.NamedList;
