+++ solr/contrib/blob-repository/src/java/org/apache/solr/s3/S3Exception.java
@@ -14,23 +14,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.blob.s3;
+package org.apache.solr.s3;
+
+import java.io.IOException;
 
 /**
- * Specific exception thrown from {@link S3StorageClient}s when the resource requested is not found (e.g., attempting
- * to fetch a key in S3 that doesn't exist).
+ * Generic exception for blob-related failures. Could originate from the
+ * {@link S3BackupRepository} or from its underlying {@link S3StorageClient}.
  */
-public class S3NotFoundException extends S3Exception {
+public class S3Exception extends IOException {
 
-    public S3NotFoundException(Throwable cause) {
+    public S3Exception(Throwable cause) {
         super(cause);
     }
 
-    public S3NotFoundException(String message) {
+    public S3Exception(String message) {
         super(message);
     }
 
-    public S3NotFoundException(String message, Throwable cause) {
+    public S3Exception(String message, Throwable cause) {
         super(message, cause);
     }
 }
