+++ solr/contrib/blob-repository/src/test/org/apache/solr/s3/S3PathsTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.blob.s3;
+package org.apache.solr.s3;
 
 import com.google.common.collect.Sets;
 import org.junit.Test;
