+++ solr/contrib/blob-repository/src/test/org/apache/solr/s3/S3ReadWriteTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.blob.s3;
+package org.apache.solr.s3;
 
 import org.apache.commons.io.IOUtils;
 import org.junit.Test;
